moved functions to utils/formatter
