+++ src/helpers/cover.js
@@ -1,28 +1,9 @@
-import {
-  parseBytes32String,
-  formatBytes32String,
-} from "@ethersproject/strings";
+import { parseBytes32String } from "@ethersproject/strings";
 
 export const getCoverImgSrc = ({ key } = { key: "" }) => {
   try {
     return `/images/covers/${parseBytes32String(key)}.svg`;
   } catch (error) {
     return `/images/covers/empty.svg`;
-  }
-};
-
-export const safeParseBytes32String = (bytes32String) => {
-  try {
-    return parseBytes32String(bytes32String);
-  } catch (error) {
-    return bytes32String;
-  }
-};
-
-export const safeFormatBytes32String = (str) => {
-  try {
-    return formatBytes32String(str);
-  } catch (error) {
-    return str;
   }
 };
