+++ src/modules/reporting/active/ActiveReportingEmptyState.jsx
@@ -4,9 +4,10 @@ import { useEffect, useState } from "react";
 import { ReportingDropdown } from "@/src/modules/reporting/reporting-dropdown";
 import { useRouter } from "next/router";
 import { actions } from "@/src/config/cover/actions";
-import { getCoverImgSrc, safeParseBytes32String } from "@/src/helpers/cover";
+import { getCoverImgSrc } from "@/src/helpers/cover";
 import { useCovers } from "@/src/context/Covers";
 import { t, Trans } from "@lingui/macro";
+import { safeParseBytes32String } from "@/utils/formatter/bytes32String";
 
 export const ActiveReportingEmptyState = () => {
   const router = useRouter();
