+++ src/modules/reporting/resolved/resolved.jsx
@@ -6,7 +6,6 @@ import { SearchAndSortBar } from "@/common/SearchAndSortBar";
 import { ResolvedReportingCard } from "@/src/modules/reporting/resolved/ResolvedReportingCard";
 import { ReportStatus } from "@/src/config/constants";
 import { useCovers } from "@/src/context/Covers";
-import { safeParseBytes32String } from "@/src/helpers/cover";
 import { useResolvedReportings } from "@/src/hooks/useResolvedReportings";
 import { useSearchResults } from "@/src/hooks/useSearchResults";
 import Link from "next/link";
@@ -14,6 +13,7 @@ import { sortData } from "@/utils/sorting";
 import { CardSkeleton } from "@/common/Skeleton/CardSkeleton";
 import { COVERS_PER_PAGE } from "@/src/config/constants";
 import { Trans } from "@lingui/macro";
+import { safeParseBytes32String } from "@/utils/formatter/bytes32String";
 
 export const ReportingResolvedPage = () => {
   const { data, loading, hasMore, handleShowMore } = useResolvedReportings();
