+++ src/pages/cover/[cover_id]/add-liquidity/index.jsx
@@ -4,9 +4,9 @@ import { CoverAddLiquidityDetailsPage } from "@/src/modules/cover/add-liquidity"
 import { ComingSoon } from "@/common/ComingSoon";
 import { isFeatureEnabled } from "@/src/config/environment";
 import { LiquidityFormsProvider } from "@/common/LiquidityForms/LiquidityFormsContext";
-import { safeFormatBytes32String } from "@/src/helpers/cover";
 import { useRouter } from "next/router";
 import { CoverInfoProvider } from "@/common/Cover/CoverInfoContext";
+import { safeFormatBytes32String } from "@/utils/formatter/bytes32String";
 
 export function getServerSideProps() {
   return {
