+++ src/pages/my-liquidity/[cover_id]/index.jsx
@@ -4,8 +4,8 @@ import { ComingSoon } from "@/common/ComingSoon";
 import { isFeatureEnabled } from "@/src/config/environment";
 import { LiquidityFormsProvider } from "@/common/LiquidityForms/LiquidityFormsContext";
 import { useRouter } from "next/router";
-import { safeFormatBytes32String } from "@/src/helpers/cover";
 import { CoverInfoProvider } from "@/common/Cover/CoverInfoContext";
+import { safeFormatBytes32String } from "@/utils/formatter/bytes32String";
 
 export function getServerSideProps() {
   return {
