+++ src/pages/reporting/[id]/new.jsx
@@ -2,7 +2,6 @@ import { useCoverInfo } from "@/src/hooks/useCoverInfo";
 import { CoverReportingRules } from "@/src/modules/reporting/CoverReportingRules";
 import { NewIncidentReportForm } from "@/src/modules/reporting/NewIncidentReportForm";
 import { ReportingHero } from "@/src/modules/reporting/ReportingHero";
-import { safeFormatBytes32String } from "@/src/helpers/cover";
 import Head from "next/head";
 import { useRouter } from "next/router";
 import { useEffect, useState } from "react";
@@ -10,6 +9,7 @@ import { useFetchCoverActiveReportings } from "@/src/hooks/useFetchCoverActiveRe
 import { ComingSoon } from "@/common/ComingSoon";
 import { isFeatureEnabled } from "@/src/config/environment";
 import { CoverInfoProvider } from "@/common/Cover/CoverInfoContext";
+import { safeFormatBytes32String } from "@/utils/formatter/bytes32String";
 
 export function getServerSideProps() {
   return {
